Add support for the from_method in the abstractor_abstraction_sources table to return a complex object.
Most of the time a user of Abstractor will simply
want to define a data point to abstract against
a textual column defined by the from_method
column in abstractor_abstraction_sources.  But
sometimes things might need to be more complex.
The specific example that drove this was trying
to abstract cancer diagnosis for pathology
specimens.  The textual fragment relevant for
abstracting a cancer diagnosis for a pathology
specimen was variable.  Could come from multiple
different sources that could only be determined
at run time.

So now the from_method is expected to return
a string or an array of hashes -- each hash expected
to follow this format:

{ source_type: ?, source_id: ? , source_method: ? }

Abstractor::AbstractorAbstractionSource#normalize_from_method_to_sources
was added to normalize the from_method to an array
of such hashes.  The various methods in AbstractorSubject
now call this method to normalize every from_method
to an array of such hashes.