Add bridge.jar file from stanford-core-nlp gem distribution of stanford-core-nlp distribution for installation rake task abstractor:setup:stanford_core_nlp From the author of the stanford-core-nlp gem
* https://github.com/louismullie/stanford-core-nlp
* http://louismullie.com/treat/stanford-core-nlp-minimal.zip
* http://louismullie.com/treat/stanford-core-nlp-full.zip

Being done for the temporary use
of a fork of stanford-core-nlp gem at 
https://github.com/mgurley/stanford-core-nlp 
to support setting custom options on Annotators 
Th current version of the stanford-core-nlp gem
does not support the setting of custom options
on Annotators.
The the latest version of stanford-core-nlp
(version 3.3.1)
has a Annotator option
ssplit.newlineIsSentenceBreak.  See
http://nlp.stanford.edu/software/corenlp.shtml
This option is not present in the
distribution of the Stanford nlp library 
provided by louismullie.