Improve the test coverage of supporting a dynamic list abstractor object type.
References #16

Querying abstractions that 'need review' or are
'reviewed' should be based on whether an
abstraction has been answered.

Closes #19

Add support for selecting indirect sources as
evidence of an abstraction.

Closes #18

Bumping to version 2.0 because this
commit includes a migration that
* adds a column abstractor_abstraction_source_type_id
to abstractor_abstraction_sources, which is a
foreign key to a new table
abstractor_abstraction_source_types.  There
are kinds of types: 'nlp suggestion',
'custom suggestion' and 'indirect'.

* moves abstractor_rule_type_id from
abstractor_subjects to
abstractor_abstraction_sources and now
confines its use to sources of only
type 'nlp suggestion'